# Use bit twiddling to speed up JSON generation.