@byroot 2025-01-29
Relative gains on M3 compared to master on the macro benchmarks:

```
== Encoding activitypub.json (52595 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after     2.529k i/100ms
Calculating -------------------------------------
               after     25.594k (± 0.9%) i/s   (39.07 μs/i) -    128.979k in   5.039874s

Comparison:
              before:    22137.7 i/s
               after:    25594.0 i/s - 1.16x  faster


== Encoding citm_catalog.json (500298 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after   135.000 i/100ms
Calculating -------------------------------------
               after      1.365k (± 0.4%) i/s  (732.82 μs/i) -      6.885k in   5.045549s

Comparison:
              before:     1371.5 i/s
               after:     1364.6 i/s - same-ish: difference falls within error


== Encoding twitter.json (466906 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +YJIT +PRISM [arm64-darwin23]
Warming up --------------------------------------
               after   266.000 i/100ms
Calculating -------------------------------------
               after      2.679k (± 0.6%) i/s  (373.22 μs/i) -     13.566k in   5.063340s

Comparison:
              before:     2379.6 i/s
               after:     2679.4 i/s - 1.13x  faster
```