@byroot 2025-01-29
This one is interesting, as it doesn't require any of the annoying feature detection SIMD impose.

But if we end up going with SIMD anyway, might as well not bother with this, right?