@byroot 2025-01-29
> It's nice to have a pure C implementation that doesn't require any special instructions.

True. I guess my only real reservation with this PR (and also with the SIMD ones) is the huge `PROCESS_BYTE` macro.

I haven't looked too much into it, but I'd really like if such huge macro wasn't necessary. So I need to take some time to experiment with some refactoring.