@byroot 2025-01-29
> I have not tested on any other than my M1 and Intel-based Laptop.

It's likely enough. x86 alone is likely 95% of Ruby usage if not more, we're probably super close to 100% if you add ARM. For other platform correctness is sufficient.