@samyron 2025-01-29
> > It's nice to have a pure C implementation that doesn't require any special instructions.
> 
> True. I guess my only real reservation with this PR (and also with the SIMD ones) is the huge `PROCESS_BYTE` macro.
> 
> I haven't looked too much into it, but I'd really like if such huge macro wasn't necessary. So I need to take some time to experiment with some refactoring.

It's not necessary. It's the existing conditional. I just didn't want to copy and paste it multiple times. 

```
if (RB_UNLIKELY(ch_len)) {
  switch(ch_len) {
  ...
  }
} else {
  pos++
}
```