@byroot 2025-01-29
> It's not necessary. It's the existing conditional. I just didn't want to copy and paste it multiple times.

Yes, I mean not having that big macro without copy-pasting either.

What I have in mind right now, but I don't know if it's really possible, would be to move the "search" part in another function, and let it having some state with a stack allocated struct so it can resume. Something very much like https://lemire.me/blog/2024/07/20/scan-html-even-faster-with-simd-instructions-c-and-c/

So the pseudo-code would look like:

```ruby
scanner_state state = {0};
while (ptr = scan(&state, ptr)) {
  // process one byte
  ptr++;
}
```

This way all the aligment consideration and such are moved in that `scan` function, and it would become the natural place where to use SIMD etc.