@byroot 2025-04-28
Is this really worth it now that we merged SIMD code? I suspect not?

I'll reopen if you think it does.